% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rand_adsl}
\alias{rand_adsl}
\alias{rand_adae}
\title{CDISC Random ADaM Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 45 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 515 rows and 82 columns.
}
\usage{
rand_adsl

rand_adae
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Random ADaM data created from \code{random.cdisc.data} package.
}
\keyword{datasets}
