% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{low1}
\alias{low1}
\alias{high1}
\title{Example Datasets with low and high dropout}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 961 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 831 rows and 6 columns.
}
\usage{
low1

high1
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These two datasets are from Mallinckrodt et al.
(https://journals.sagepub.com/doi/pdf/10.1177/2168479013501310) which is available
via (https://www.lshtm.ac.uk/research/centres-projects-groups/missing-data#dia-missing-data).
The original data were from 2 nearly identically designed antidepressant clinical
trials that were originally reported by Goldstein et al and Detke et al.
Assessments on the Hamilton 17-item rating scale for depression (HAMD17)19 were
taken at baseline and weeks 1, 2, 4, 6, and 8 in each trial.
}
\keyword{datasets}
