% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{prop_odds_ratio}
\alias{prop_odds_ratio}
\alias{s_propci}
\alias{h_prep_prop}
\title{Computing Proportion and Odds Ratio}
\usage{
s_propci(
  data,
  var,
  by = NULL,
  by.level = NULL,
  event = NULL,
  conf.level = 0.95,
  method = c("clopper-pearson", "wald", "waldcc", "wilson", "wilsoncc",
    "modified wilson", "jeffreys", "modified jeffreys", "agresti-coull"),
  diff.method = c("wald", "waldcc", "score", "scorecc", "mn", "mee", "blj", "ha", "hal",
    "jp"),
  alternative = c("two.sided", "less", "greater"),
  ...
)

h_prep_prop(data, var, by, by.level, event)
}
\arguments{
\item{data}{(\code{data.frame})\cr a data frame as input.}

\item{var}{(\code{string})\cr target variable name for estimation.}

\item{by}{(\code{string})\cr a optional variable to group by. If null, use the whole data.}

\item{by.level}{(\code{vector})\cr an optional vector for encoding \code{var} as a factor
and the second level will be as the reference group. If null, use the default
order to encode.}

\item{event}{(\code{numeric} or \code{character})\cr an option to define which one as the
event in the elements of \code{var}. By default, the positive and maximal one if
the \code{var} variable is numeric, or the first one of if the \code{var} variable is
character/factor.}

\item{conf.level}{(\code{numeric})\cr significance level for the returned confidence
interval.}

\item{method}{(\code{string})\cr string specifying which method to calculate the
confidence interval for binomial proportions, default is Clopper-Pearson. More
details see \link[DescTools:BinomCI]{DescTools::BinomCI}.}

\item{diff.method}{(\code{string})\cr string specifying which method to calculate the
confidence interval for difference between binomial proportions, default is
Wald. More details see \link[DescTools:BinomDiffCI]{DescTools::BinomDiffCI}.}

\item{alternative}{(\code{string})\cr string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{...}{other arguments to be passed to \link[DescTools:BinomCI]{DescTools::BinomCI}.}
}
\value{
\itemize{
\item \code{s_propci} returns an object of class \code{prop_ci} that is a list contains
proportion rate, proportion difference and arguments.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute confidence interval for proportion and difference of binomial response,
and odds ratio using the \code{DescTools} package. As an alternative, use \code{stats::glm}
with \code{logit} link to estimate the odds ratio. Regarding to stratified odds ratio,
use Cochran-Mantel-Haenszel test (\code{stats::mantelhaen.test}) or conditional logistic
regression (\code{survival::clogit}) to handle with.
}
\section{Functions}{
\itemize{
\item \code{s_propci()}: Computes the confidence interval of proportion/response
rate and difference of two proportions/rates.

\item \code{h_prep_prop()}: Helper Function for Pre-processing Proportion Data.

}}
\note{
The proportion difference can only support the two-levels for \code{by} variable.
}
\examples{
set.seed(12)
dta <- data.frame(
  orr = sample(c(1, 0), 100, TRUE),
  trtp = factor(rep(c("TRT", "PBO"), each = 50)),
  strata1 = factor(sample(c("A", "B"), 100, TRUE)),
  strata2 = factor(sample(c("C", "D"), 100, TRUE)),
  strata3 = factor(sample(c("E", "F"), 100, TRUE))
)

# Not having by variable:
s_propci(dta, var = "orr")

# Two levels of by variable:
s_propci(dta, var = "orr", by = "trtp")
# opposite order of trtp with above:
s_propci(dta, var = "orr", by = "trtp", by.level = c("TRT", "PBO"), event = 1)
}
