% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{rrPostProb}
\alias{rrPostProb}
\title{Simulation of Sample Size Determination by Bayesian}
\usage{
rrPostProb(nsample, cutoff, p, a = 1, b = 1, succs = NULL, nrep = 1e+05)
}
\arguments{
\item{nsample}{(\code{integer})\cr integer value of specified sample size.}

\item{cutoff}{(\code{numeric})\cr numeric cutoff of response rate.}

\item{p}{(\code{numeric})\cr numeric vector of true response rate.}

\item{a, b}{(\code{numeric})\cr parameters of Bayesian priors, defaults are a=1 and
b=1 which is a weekly informative prior.}

\item{succs}{(\code{numeric})\cr a threshold value to define pass for each simulation.}

\item{nrep}{(\code{integer})\cr times of simulation.}
}
\value{
a vector of probability for each true response rate.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Simulation to obtain trial operational characteristic
}
\details{
The Prior is P(p)~Beta(1,1), likelihood is P(k|p)~Binomial(n,p), and posterior
can be P(p|k)~Bate(k+1, n-k+1).
}
\examples{
# Given that the operational characteristics are based on at least 70\% posterior
# confidence that the true rate exceeds a threshold of interest, what is the
# chance of confirming a true response of at least 75\% given the assumptions
# of various sample sizes and true response rates(0.75, 0.8, 0.85, 0.9)?
# It should be noted that operational characteristics are based on a Bayesian
# model without considering baseline stratification factors.
rrPostProb(nsample = 66, cutoff = 0.75, p = c(0.75, 0.8, 0.85, 0.9), succs = 0.7)
}
