% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-survival.R, R/pkg-methods.R
\name{s_get_coxph}
\alias{s_get_coxph}
\alias{print.s_coxph}
\title{Summarize Cox Proportional Hazards Regression Model}
\usage{
s_get_coxph(
  data,
  formula,
  ties = c("efron", "breslow", "exact"),
  conf_level = 0.95,
  strata = NULL,
  pval_method = c("all", "log", "sc", "wald"),
  pairwise = FALSE,
  ...
)

\method{print}{s_coxph}(x, export_pdf = FALSE, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr a data frame as input.}

\item{formula}{(\code{formula})\cr a formula with survival object.}

\item{ties}{(\code{string})\cr string specifying the the method for tie handling,
default is efron. See more details in \code{\link[survival:coxph]{survival::coxph()}}.}

\item{conf_level}{(\code{numeric})\cr significance level for a two-side confidence
interval on survival curve, default is 0.95.}

\item{strata}{(\code{character})\cr a character vector used for stratification.}

\item{pval_method}{(\code{string})\cr string specifying the the method for tie handling,
default is to present all three methods (Likelihood ratio test, Wald test and
Score (logrank) test).}

\item{pairwise}{(\code{logical})\cr whether to conduct the pairwise comparison.}

\item{...}{other arguments to be passed to \code{\link[survival:coxph]{survival::coxph()}}.}

\item{export_pdf}{(\code{logical})\cr whether to export PDF to the current directory.}
}
\value{
An object of class \code{s_coxph} is a list contains hazards ratio and p-value tables.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function summarizes \code{coxph} model with or without stratification, and provides
hazards ratio, confidence interval and p-value for common clinical survival analysis.
}
\section{Functions}{
\itemize{
\item \code{print(s_coxph)}: prints survival analysis summary from \code{coxph}.

}}
\examples{
data("whas500")

# reorder the grouping variable
dat <- whas500 |>
  dplyr::mutate(
    AFB = factor(AFB, levels = c(1, 0))
  )
s_get_coxph(data = dat, formula = Surv(LENFOL, FSTAT) ~ AFB)

s_get_coxph(data = dat, formula = Surv(LENFOL, FSTAT) ~ AFB, pval_method = "sc")

# specify the stratified analysis
s_get_coxph(
  data = dat,
  formula = Surv(LENFOL, FSTAT) ~ AFB,
  strata = c("AGE", "GENDER")
)

# dummy three groups
set.seed(123)
subj <- sample(dat$ID, 100)
dat2 <- whas500 |>
  dplyr::mutate(
    AFB = dplyr::case_when(
      ID \%in\% subj ~ 2,
      TRUE ~ AFB
    ),
    AFB = factor(AFB, levels = c(1, 2, 0))
  )
s_get_coxph(data = dat2, formula = Surv(LENFOL, FSTAT) ~ AFB)
}
\references{
SAS code for your reference with consistent results.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ods output hazardratios=hrs;
ods output ParameterEstimates=hr_est;
ods listing close;
proc phreg data=whas500 alpha=0.05;
  class AFB (param=ref ref='1') AGE GENDER;
  model LENFOL*FSTAT(0)=AFB / ties=efron risklimits=wald;
  /* 	strata AGE GENDER; */
  hazardratio AFB / diff=ref;
quit;
ods listing;
}\if{html}{\out{</div>}}
}
\keyword{internal}
