% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-lsmeans.R, R/pkg-methods.R
\name{s_get_lsmeans}
\alias{s_get_lsmeans}
\alias{print.s_lsmeans}
\title{Summarize Least-squares Means from Models}
\usage{
s_get_lsmeans(
  object,
  var,
  by = NULL,
  contrast = TRUE,
  null = 0,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater"),
  ...
)

\method{print}{s_lsmeans}(x, ...)
}
\arguments{
\item{object}{(\verb{fitted model})\cr any fitted model that can be accepted by
\code{emmeans::emmeans()}, such as object from \code{lm} for ANCOVA and \code{mmrm} for MMRM.}

\item{var}{(\code{string})\cr string specifying the name of the predictor, such as
it would be treatment group variable (TRT01PN) in the clinical trials. Default
the first level is defined as reference group.}

\item{by}{(\code{string})\cr string specifying the name of the grouped variable.
The estimates and contrasts will be evaluated separately for each elements
from the grouped variable.}

\item{contrast}{(\code{logical})\cr whether to perform the contrasts, pairwise
comparisons (required usage) and hypothesis testing or not, default is TRUE.}

\item{null}{(\code{numeric})\cr null hypothesis value, it will also be referred to
the margin of superiority and non-inferiority design, default is 0.}

\item{conf.level}{(\code{numeric})\cr significance level for the returned confidence
interval and hypothesis testing, default is 0.95.}

\item{alternative}{(\code{string})\cr string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". See the special
section below for more details.}

\item{...}{other arguments to be passed to \code{\link[emmeans:emmeans]{emmeans::emmeans()}}.}
}
\value{
An object of class \code{s_lsmeans} is a list contains several summary tables and
fit model as following components:
\itemize{
\item \code{model} information of fitted model.
\item \code{lsm_est} estimate \code{tibble} table of Ls-means with corresponding statistics.
If \code{by} parameter is specified, all estimates for each \code{var} and \code{by} will
be presented.
\item \code{lsm_contr} contrast \code{tibble} table with pairwise comparisons that contains
corresponding statistics, if the \code{contrast} is \code{TRUE}.
\item \code{params} essential parameters.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

To estimate the Least-squares Means along with corresponding statistics and
confidence interval, and perform the hypothesis testing, using the \code{emmeans}
package.
}
\details{
For example, when the null hypothesis is θ equal to \code{0}, the \code{side} should be
\code{two.sided}, and \code{null} is defined as \code{0}. For the non-inferiority trial,
such as when the null hypothesis is θ inferiority to \code{-2}, the \code{side} should
be \code{greater}, and \code{null} is defined as \code{-2}. For the superiority trial,
when the null hypothesis is θ not superiority to \code{2}, the \code{side} should be
\code{greater}, and \code{null} is defined as \code{2}.
}
\section{Functions}{
\itemize{
\item \code{print(s_lsmeans)}: prints Least-squares Means summary.

}}
\note{
There is no any p value adjusted method involved.
}
\examples{
# refactor the level order:
data(fev_data, package = "mmrm")

# fit MMRM model:
fit <- mmrm::mmrm(
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  reml = TRUE, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear",
  data = fev_data
)

# estimate ARMCD by visits:
s_get_lsmeans(fit, "ARMCD", by = "AVISIT")

# estimate ARMCD by visits for superiority testing with null hypothesis of 2:
s_get_lsmeans(fit, "ARMCD", by = "AVISIT", null = 2, alternative = "greater")

# fit ANCOVA model:
fit2 <- fev_data |>
  dplyr::filter(VISITN == 4 & !is.na(FEV1)) |>
  lm(formula = FEV1 ~ FEV1_BL + RACE + SEX + ARMCD)

s_get_lsmeans(fit2, "ARMCD")
}
\references{
SAS code for your reference with consistent results.

\if{html}{\out{<div class="sourceCode">}}\preformatted{proc mixed data=fev_data;
  class ARMCD(ref='PBO') AVISIT RACE SEX USUBJID;
  model FEV1 = RACE SEX ARMCD ARMCD*AVISIT / ddfm=KR;
  repeated AVISIT / subject=USUBJID type=UN r rcorr;
  lsmeans ARMCD*AVISIT / cl alpha=0.05 diff slice=AVISIT;
  lsmeans ARMCD / cl alpha=0.05 diff;
  lsmestimate ARMCD*AVISIT [1,1 4] [-1,2 4] / cl upper alpha=0.025 testvalue=2;
  ods output lsmeans=lsm diffs=diff LSMEstimates=est;
run;
}\if{html}{\out{</div>}}
}
\keyword{internal}
