% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{stabiot-package}
\alias{stabiot}
\alias{stabiot-package}
\title{\code{stabiot} Package}
\description{
\code{stabiot} Common Statistical Analysis for Clinical Trials in Biotech.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaigu1990/stabiot}
  \item Report bugs at \url{https://github.com/kaigu1990/stabiot/issues}
}

}
\author{
\strong{Maintainer}: Kai Gu \email{gukai1212@163.com} [copyright holder]

}
