% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whas500}
\alias{whas500}
\title{Worcester Heart Attack Study Data}
\format{
This data set contains 500 subjects and 19 variables.
\itemize{
\item ID:     Identification Code               (1 - 500)
\item AGE:    Age at Hospital Admission         (Years)
\item GENDER: Gender                            (0 = Male, 1 = Female)
\item HR:     Initial Heart Rate                (Beats per minute)
\item SYSBP:  Initial Systolic Blood Pressure   (mmHg)
\item DIASBP: Initial Diastolic Blood Pressure  (mmHg)
\item BMI:    Body Mass Index                   (kg/m^2)
\item CVD:    History of Cardiovascular Disease (0 = No, 1 = Yes)
\item AFB:    Atrial Fibrillation               (0 = No, 1 = Yes)
\item SHO:    Cardiogenic Shock                 (0 = No, 1 = Yes)
\item CHF:    Congestive Heart Complications    (0 = No, 1 = Yes)
\item AV3:    Complete Heart Block              (0 = No, 1 = Yes)
\item MIORD:  MI Order                          (0 = First, 1 = Recurrent)
\item MITYPE: MI Type                           (0 = non Q-wave, 1 = Q-wave)
\item YEAR:   Cohort Year                       (1 = 1997, 2 = 1999, 3 = 2001)
\item LOS:    Length of Hospital Stay           (Days between Hospital Discharge and Hospital Admission)
\item DSTAT:  Discharge Status from Hospital    (0 = Alive, 1 = Dead)
\item LENFOL: Total Length of Follow-up         (Days between Date of Last Follow-up and Hospital Admission Date)
\item FSTAT:  Vital Status at Last Follow-up    (0 = Alive 1 = Dead)
}
}
\usage{
whas500
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a data of 500 subjects of the Worcester Heart Attack study (also called WHAS500)
that demonstrates several factors, such as age, gender and BMI, that may influence
survival time after heart attack. Follow up time for all participants begins at
the time of hospital admission after heart attack and ends with death or loss
to follow up (censoring).
}
\references{
Hosmer, D.W. and Lemeshow, S. and May, S. (2008) Applied Survival
Analysis: Regression Modeling of Time to Event Data: Second Edition,
John Wiley and Sons Inc., New York, NY.
}
\keyword{datasets}
