% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R, R/pkg-methods.R, R/utils.R
\name{prop_odds_ratio}
\alias{prop_odds_ratio}
\alias{print.prop_ci}
\alias{print.or_ci}
\alias{s_propci}
\alias{s_odds_ratio}
\alias{h_prep_prop}
\title{Computing Proportion and Odds Ratio}
\usage{
\method{print}{prop_ci}(x, ...)

\method{print}{or_ci}(x, ...)

s_propci(
  data,
  var,
  by = NULL,
  by.level = NULL,
  resp = NULL,
  conf.level = 0.95,
  method = c("clopper-pearson", "wald", "waldcc", "wilson", "wilsoncc",
    "modified wilson", "jeffreys", "modified jeffreys", "agresti-coull"),
  diff.method = c("wald", "waldcc", "score", "scorecc", "mn", "mee", "blj", "ha", "hal",
    "jp"),
  alternative = c("two.sided", "less", "greater"),
  ...
)

s_odds_ratio(
  data,
  var,
  by,
  by.level = NULL,
  resp = NULL,
  strata = NULL,
  conf.level = 0.95,
  or.glm = FALSE,
  or.method = c("wald", "mle", "midp"),
  strata.method = c("CMH", "clogit"),
  correct = FALSE,
  exact = FALSE
)

h_prep_prop(data, var, by, by.level, resp)
}
\arguments{
\item{...}{other arguments to be passed to \link[DescTools:BinomCI]{DescTools::BinomCI}.}

\item{data}{(\code{data.frame})\cr a data frame as input.}

\item{var}{(\code{string})\cr target variable name for estimation.}

\item{by}{(\code{string})\cr an optional variable to group by. If null, use the whole data.}

\item{by.level}{(\code{vector})\cr an optional vector for encoding \code{var} as a factor
and the first level will be as the reference group. If null, use the default
order to encode.}

\item{resp}{(\code{numeric} or \code{character})\cr an option to define which one as the
respondor in the elements of \code{var}. By default, the positive and maximal one if
the \code{var} variable is numeric, or the first one of if the \code{var} variable is
character/factor.}

\item{conf.level}{(\code{numeric})\cr significance level for the returned confidence
interval.}

\item{method}{(\code{string})\cr string specifying which method to calculate the
confidence interval for binomial proportions, default is Clopper-Pearson. More
details see \link[DescTools:BinomCI]{DescTools::BinomCI}.}

\item{diff.method}{(\code{string})\cr string specifying which method to calculate the
confidence interval for difference between binomial proportions, default is
Wald. More details see \link[DescTools:BinomDiffCI]{DescTools::BinomDiffCI}.}

\item{alternative}{(\code{string})\cr string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{strata}{(\code{vector})\cr a character vector is used for stratification.}

\item{or.glm}{(\code{logical})\cr a logical indicating whether to use logit method to
calculate the odds ratio. If TRUE, the \code{glm} with logit is used, otherwise the
common method from \code{DescTools::OddsRatio}.}

\item{or.method}{(\code{string})\cr string indicating the method from \code{DescTools::OddsRatio}
used to calculate odds ratio when the \code{or.glm} is FALSE.}

\item{strata.method}{(\code{string})\cr string indicating the method used to calculate
stratified odds ratio, Cochran-Mantel-Haenszel Chi-Squared test (CMH) or conditional
logistic regression (clogit).}

\item{correct}{(\code{logical})\cr a logical indicating whether to apply continuity
correction when computing the test statistic.}

\item{exact}{(\code{logical})\cr a logical indicating whether the Mantel-Haenszel
test or the exact conditional test (given the strata margins) should be computed.}
}
\value{
\itemize{
\item \code{s_propci} returns an object of class \code{prop_ci} that is a list contains
proportion rate, proportion difference and input arguments.
}

\itemize{
\item \code{s_odds_ratio} returns an object of class \code{or_ci} that is a list contains
odds ratio with or without stratification and input arguments.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute confidence interval for proportion and difference of binomial response,
and odds ratio using the \code{DescTools} package. As an alternative, use \code{stats::glm}
with \code{logit} link to estimate the odds ratio. Regarding to stratified odds ratio,
use Cochran-Mantel-Haenszel test (\code{stats::mantelhaen.test}) or conditional logistic
regression (\code{survival::clogit}) to handle with. However this CMH test function is
limited to the 2 by 2 by K scenario, whereas conditional logistic regression
can handle 2 by N by K case.
}
\section{Functions}{
\itemize{
\item \code{print(prop_ci)}: prints proportion and confidence interval.

\item \code{print(or_ci)}: prints odds ratio and confidence interval.

\item \code{s_propci()}: Computes the confidence interval of proportion/response
rate and difference of two proportions/rates.

\item \code{s_odds_ratio()}: Computes the confidence interval of common odds ratio.
And also provides stratified odds ratio with Cochran-Mantel-Haenszel chi-squared
test and conditional logistic regression.

\item \code{h_prep_prop()}: Helper Function for Pre-processing Proportion Data.

}}
\note{
\itemize{
\item The proportion difference can only support the two-levels for \code{by} variable.
\item Compute odds ratio when \code{or.glm} is FALSE, 0.5 will be added to 2 x 2 table
in case of zero entries. More details can be found in \link[DescTools:OddsRatio]{DescTools::OddsRatio}.
\item The arguments of this function can only match few parameters of SAS \verb{proc freq}
to calculate the odds ratio. For example, \code{mantelhaen.test} corresponds the
Mantel-Haenszel Estimator with the same reference article, and using \code{correct = FALSE}
can produce the Wald confidence interval same as the default parameters of SAS.
}
}
\examples{
set.seed(12)
dta <- data.frame(
  orr = sample(c(1, 0), 150, TRUE),
  trtp = factor(rep(c("TRT1", "TRT2", "PBO"), each = 50)),
  strata1 = factor(sample(c("A", "B"), 150, TRUE)),
  strata2 = factor(sample(c("C", "D"), 150, TRUE)),
  strata3 = factor(sample(c("E", "F"), 150, TRUE))
)

# not having by variable:
s_propci(dta, var = "orr")

# two levels of by variable:
s_propci(dta, var = "orr", by = "trtp")
# opposite order of trtp with above:
s_propci(dta, var = "orr", by = "trtp", by.level = c("TRT1", "TRT2", "PBO"), resp = 1)

# without stratification:
s_odds_ratio(dta, var = "orr", by = "trtp", or.method = "wald")

# with three stratifications, strata1 - strata3:
s_odds_ratio(
  dta,
  var = "orr", by = "trtp",
  strata = c("strata1", "strata2", "strata3"),
  strata.method = "CMH",
  correct = FALSE
)

# using condition logistic regression:
s_odds_ratio(
  dta,
  var = "orr", by = "trtp",
  or.glm = TRUE,
  strata = c("strata1", "strata2", "strata3"),
  strata.method = "clogit"
)
}
\references{
SAS code for your reference with consistent results.
\itemize{
\item The rate and difference of proportions are estimated:
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{proc means data=dta noprint;
  class trtp orr strata1 strata2 strata3;
  types trtp*orr*strata1*strata2*strata3;
  output out=mat(drop=_type_ rename=_freq_=count);
run;

ods listing close;
proc freq data=mat;
  by trtp;
  weight count/zeros;
  tables orr /binomial(level="1") alpha=0.05;
  exact binomial;
  ods output binomial=orrci;
run;

proc freq data=mat(where=(trtp in ('PBO' 'TRT1')));
  weight count/zeros;
  tables trtp*orr /riskdiff(CL=wald column=1) alpha=0.05;
  ods output PdiffCLs=diff;
  run;
ods listing;
}\if{html}{\out{</div>}}

\itemize{
\item The common odds ratio and stratified odds ratio with CMH test and conditional logistic are estimated:
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{ods listing close;
  proc freq data=mat(where=(trtp in ('PBO' 'TRT1')));
  weight count/zeros;
  tables trtp*orr / relrisk chisq;
  ods output ChiSq=pval RelativeRisks=ci;
run;

proc freq data=mat(where=(trtp in ('PBO' 'TRT1')));
  weight count/zeros;
  tables strata1*strata2*strata3*trtp*orr / relrisk cmh;
  ods output cmh=cmhpval CommonRelRisks=cmhci;
run;

proc logistic data=mat;
  freq count;
  class trtp strata1 strata2 strata3 / param=ref ref=first;
  strata strata1 strata2 strata3;
  model orr(event='1')=trtp;
run;
ods listing;
}\if{html}{\out{</div>}}
}
\keyword{internal}
